Include Headers in PublishFile task and remove separate PublishFilesWithHeaders classes (task still provided for backwards compatibility)
